[refactor] : main, app 파일을 src 밑으로 넣기
