+++ index.html
@@ -7,6 +7,6 @@
   </head>
   <body>
     <main id="app"></main>
-    <script type="module" src="/main.js"></script>
+    <script type="module" src="/src/main.js"></script>
   </body>
 </html>
