+++ src/main.js
@@ -1,6 +1,6 @@
 import App from "./App.js";
-import { SELECTOR } from "./src/constants/selector.js";
-import { $ } from "./src/utils/dom.js";
+import { SELECTOR } from "./constants/selector.js";
+import { $ } from "./utils/dom.js";
 
 const $app = $(`#${SELECTOR.ROOT}`);
